Fix build of kernel
The installRaw function, whilst outputting the kernel in a way that QEMU
handles it properly, doesn't work for gdb, objdump, etc.

I've realised that .addArtifactArg exists, which removes this bodge
entirely, while also removing the need to add an explicit dependency, or
worry about what path we're in.

Also, pull out the number of cores to a constant (could be converted to
a build argument?).