Flattened device tree blob parsing
QEMU passes in a pointer to the device tree blob in a1, so pass this
into Zig as the first argument to entry.

Then, write a simple parser that prints out the whole tree to UART. Once
drivers and memory allocation is done, we should probably store this and
pass it to drivers.