Fix fdtb comptime usage
I really didn't get how comptime worked. The whole block is evaluated at
comptime, not just the condition. By using a comptime if, I thought I
was forcing the actual branch to be optimised out. In actuality, for
const variables this occurs regardless.

As of 0.11.0 this is also a compile error, as the statements inside the
block can't be evaluated at compile time.