Attempt to remove offset function call
Calling into a function twice to get a constant offset into a struct is
kinda bleh, so this is an attempt to get zig to export that offset to
the assembly.

It currently DOESN'T WORK, because the symbl process_stack_ptr is
actually the address in memory of this constant. I'll either need to
deref this to a register and add it to the base, or reincorporate this
into zig source (using naked functions?)