Implement a StructList data structure
A doubly-linked list of structs, where each element is contained within
the struct. Like the list defined in PintOS or the Kernel linked list in
Linux.

There are some tests, but build.zig isn't configured to run tests (I'd
need to go back and either make the tests run under x86_64 or get them
to run in QEMU), but `zig test ...` works.