Create a trap stack within the Process struct
When we execute a trap handler, we need a stack for it to run on. I've
chosen not to use the user programs own stack pointer as it leads to a
whole bunch of issues (what if it spans different virtual pages or if
accessing it leads to a page fault?).

Instead, we'll create a per-Process stack that we'll switch to whenever
we trap. I didn't really have a size in mind, so I've made the whole
process struct a whole page long in total, with the stack taking up
however much is left once all the other fields are full.

I've also added a "magic" value we can check that a pointer we have to a
Process struct actually might point to a Process struct. Note that this
magic value *won't* necessarily detect a stack overflow, because Zig can
reorder the fields in a struct.