Add compile-time check for Process stack location
The trap-handling code relies on the Process' stack array being at the
end of the Process struct. Zig will rearrange struct members, so there's
a chance it will be moved around, causing bugs.

I can't think of a nice way to get around this - we can't give the
struct a specific memory ordering by making it extern (as it contains
types that aren't extern-able, the State enum and StructList elems) or
by making it a packed enum (as the arrays for saved registers and
process name don't have a guaranteed in-memory representation) - so
instead, I'll just throw a compile-time assertion that it is at the end.

If the assertion breaks, I'll at least know where the bug is.