Add permissions to paging functions
Add extra parameters for passing in the permissions a page should get
when it's created/reset. Take the opportunity to make an invalid
permission combination (such as write-only - set privileged spec) a
panic, rather than unreachable code.

We combine permissions when creating pages, but don't when we set a
mapping. This is because when creating pages, if we do try to create a
page, it's likely because we're trying to write something new to a page
that already contains some data. So, we combine permissions.