Add a test user program
The operating system's going to need programs to run, so write a simple
"Hello world" C program that outputs to UART.

We handwrite our own _start so it compiles with a Clib.

We compile it with optimisations on so stack accesses are optimised out.

Add a step to the Zig build that runs the Makefile in user/programs so
we can include binaries in the compiled Zig kernel with @embedFile

Update README with new dependancies for compiling user programs (yes, I
could just compile the user programs with Zig instead, but I have GCC so
may as well use it)