Load registers when running first program
When we launch the initial process, load its registers from the initial
process' saved array. This has highlighted a bug in the trap stub
implementation where we were setting the stack pointer, but not the
frame pointer, which led to issues when the Zig-generated assembly
surrounding the inline assembly tried to save the 'running' output.