Pass around process fault cause as well as pc
In addition to the fault address, the value of the program counter when
the fault occured, also store the fault cause in the Process struct.
This can have a couple different values, but most importantly, on an
access/page fault, it is the virtual address the user process tried to
access.

To avoid confusion, rename fault_addr to pc