Replace per-process trap stack with per-hart one
As we can only be processing one trap at a time (interrupts are disabled
whilst we're in machine mode, and exceptions only occur due to kernel
bugs), we don't need to have one trap stack for every different process.
So, remove it from the process struct and instead reuse the kernel main
stack.

I've renamed main, and reworked some of the documentation to make it
clear this occurs, so that we DO NOT rely on things put on the stack
in kinit hanging around.

I've also removed the extern *anyopaque variable for accessing the stack
end linker variable, as we only access it from inside machine code,
where a simple "la ... _stack_end" is easier.