Implement process/page table destruction
Running Process' can have their state set to .DYING by a handler, which
will cause them to be killed the next time a scheduling decision is
requested. Alternatively destroy can be called on a process that is
ready or blocked to kill it.

We can now kill a process when it page faults and stack growth wouldn't
save it. We use to terminate the hello and fib processes' when they're
done, rather than looping forever.